move entityTarget , PacketPump, and receivedPacket into proper file. ( class or structure.)
