+++ megamek/src/megamek/common/weapons/ACAPHandler.java
@@ -26,7 +26,7 @@ import megamek.common.Report;
 import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.server.Server;
-import megamek.server.DamageType;
+import megamek.common.enums.DamageType;
 
 /**
  * @author Andrew Hunter
