+++ megamek/src/megamek/common/weapons/LRMFragHandler.java
@@ -25,7 +25,7 @@ import megamek.common.TargetRoll;
 import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.server.Server;
-import megamek.server.DamageType;
+import megamek.common.enums.DamageType;
 
 /**
  * @author Sebastian Brocks
