+++ megamek/src/megamek/common/weapons/PopUpMineLauncherHandler.java
@@ -28,7 +28,7 @@ import megamek.common.Tank;
 import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.server.Server;
-import megamek.server.DamageType;
+import megamek.common.enums.DamageType;
 
 public class PopUpMineLauncherHandler extends AmmoWeaponHandler {
     private static final long serialVersionUID = -6179453250580148965L;
