+++ megamek/src/megamek/common/weapons/RifleWeaponHandler.java
@@ -28,7 +28,7 @@ import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.common.options.OptionsConstants;
 import megamek.server.Server;
-import megamek.server.DamageType;
+import megamek.common.enums.DamageType;
 
 /**
  * @author Jason Tighe
