+++ megamek/src/megamek/common/weapons/SRMInfernoHandler.java
@@ -29,7 +29,7 @@ import megamek.common.actions.WeaponAttackAction;
 import megamek.common.enums.GamePhase;
 import megamek.common.options.OptionsConstants;
 import megamek.server.Server;
-import megamek.server.DamageType;
+import megamek.common.enums.DamageType;
 
 /**
  * @author Sebastian Brocks
