+++ megamek/src/megamek/common/weapons/SRMTandemChargeHandler.java
@@ -30,7 +30,7 @@ import megamek.common.Tank;
 import megamek.common.ToHitData;
 import megamek.common.actions.WeaponAttackAction;
 import megamek.server.Server;
-import megamek.server.DamageType;
+import megamek.common.enums.DamageType;
 
 /**
  * @author Jason Tighe
