+++ megamek/src/megamek/server/ReceivedPacket.java
@@ -1,0 +1,16 @@
+package megamek.server;
+
+import megamek.common.net.Packet;
+
+public class ReceivedPacket {
+
+        public int connId;
+        public Packet packet;
+
+        ReceivedPacket(int cid, Packet p) {
+            packet = p;
+            connId = cid;
+        }
+
+
+}
